+++ src/analyze/src/views/Dashboards.vue
@@ -26,6 +26,9 @@ export default {
       'dashboards',
       'reports',
     ]),
+    isActive() {
+      return dashboard => dashboard.id === this.activeDashboard.id;
+    },
   },
   methods: {
     ...mapActions('dashboards', [
@@ -33,9 +36,6 @@ export default {
       'setDashboard',
       'getActiveDashboardReportsWithQueryResults',
     ]),
-    isActive(dashboard) {
-      return dashboard.id === this.activeDashboard.id;
-    },
     toggleNewDashboardModal() {
       this.isNewDashboardModalOpen = !this.isNewDashboardModalOpen;
     },
@@ -62,47 +62,83 @@ export default {
 
     <div class="container view-body is-fluid">
       <section>
-        <div class="columns is-gapless">
+        <div class="columns">
+
           <aside class="column is-one-quarter">
-            <nav class="panel has-background-white">
-              <p class="panel-heading">
-                Dashboards
-              </p>
+            <div class="box">
 
-              <div class='panel-block'>
-                <a class='button is-secondary is-fullwidth'
-                    @click="toggleNewDashboardModal">New Dashboard</a>
-              </div>
-
-              <div v-for="dashboard in dashboards"
-                  class='panel-block'
-                  :class="{'is-active': isActive(dashboard)}"
-                  :key="dashboard.id"
-                  @click="setDashboard(dashboard)">
-                <div>
-                  <div>{{dashboard.name}}</div>
+              <div class="columns is-vcentered">
+                <div class="column is-three-fifths">
+                  <h2 class="title is-5">Dashboards</h2>
+                </div>
+                <div class="column is-two-fifths">
+                  <div class="buttons is-right">
+                    <button
+                      class="button is-success"
+                      @click="toggleNewDashboardModal">New</button>
+                  </div>
                 </div>
               </div>
 
-            </nav>
-          </aside>
-          <div class="column is-three-quarters">
+              <div class="panel">
+                <a v-for="dashboard in dashboards"
+                    class='panel-block space-between has-text-weight-medium'
+                    :class="{'is-active': isActive(dashboard)}"
+                    :key="dashboard.id"
+                    @click="setDashboard(dashboard)">
+                  {{dashboard.name}}
+                </a>
+              </div>
 
-            <h1>{{activeDashboard.name}}</h1>
-            <h2 v-if="activeDashboard.description">{{activeDashboard.description}}</h2>
-            <hr v-if="activeDashboardReports.length">
-            <div
-              class='box'
-              v-for="report in activeDashboardReports"
-              :key="report.id">
-              <p>{{report.name}}</p>
-              <chart :chart-type='report.chartType'
-                      :results='report.queryResults'
-                      :result-aggregates='report.queryResultAggregates'></chart>
             </div>
+          </aside>
 
-            <NewDashboardModal v-if="isNewDashboardModalOpen" @close="toggleNewDashboardModal" />
+          <div class="column is-three-quarters">
+            <div class="box">
+              <template v-if='activeDashboard.name'>
+                <div class="columns is-vcentered">
+                  <div class="column">
+                    <h2 class="title is-5">{{activeDashboard.name}}</h2>
+                    <h3
+                      class='subtitle'
+                      v-if="activeDashboard.description">{{activeDashboard.description}}</h3>
+                  </div>
+                  <div
+                    v-if='activeDashboard.name'
+                    class="column">
+                    <div class="buttons is-right">
+                      <a
+                        class='button tooltip is-tooltip-warning is-tooltip-multiline is-tooltip-left'
+                        data-tooltip='This feature is queued. Feel free to contribute at gitlab.com/meltano/meltano/issues.'>Add Report</a>
+                    </div>
+                  </div>
+                </div>
+
+                <div class='content'
+                  v-for="report in activeDashboardReports"
+                  :key="report.id">
+                  <h5 class='has-text-centered'>{{report.name}}</h5>
+                  <chart :chart-type='report.chartType'
+                          :results='report.queryResults'
+                          :result-aggregates='report.queryResultAggregates'></chart>
+                </div>
+              </template>
+
+              <template v-else>
+                <div class="column content">
+                  <p>Click "New" to the left and add reports to it to view them here. You can add a report by:</p>
+                  <ul>
+                    <li>In Analyze, clicking "Add to Dashboard" after creating a report</li>
+                    <li>Clicking the "Add Report" button to an existing dashboard (coming soon)</li>
+                  </ul>
+                </div>
+              </template>
+
+              <NewDashboardModal v-if="isNewDashboardModalOpen" @close="toggleNewDashboardModal" />
+
+            </div>
           </div>
+
         </div>
       </section>
     </div>
@@ -110,5 +146,5 @@ export default {
 
 </template>
 
-<style lang="scss" scoped>
+<style lang="scss">
 </style>